マージできるようにStockUtilをリネーム
