+++ CalcLib/CalcLib.csproj
@@ -57,6 +57,7 @@
     <Compile Include="Maeda\CalcSvcMaeda.cs" />
     <Compile Include="CalcSvc.cs" />
     <Compile Include="CalcContext.cs" />
+    <Compile Include="Util\StockUtilYamamoto.cs" />
     <Compile Include="Yamamoto\ApplicationFactory.cs" />
     <Compile Include="Maeda\Util\EnumerableExt.cs" />
     <Compile Include="Maeda\Util\SvcActivator.cs" />
